refactor out the json marshaling to a common place.
